use long for text drawing

+++ TOMB4/game/text.h
@@ -3,9 +3,9 @@
 
 void InitFont();
 void UpdatePulseColour();
-long GetStringLength(const char* string, short* top, short* bottom);
-void DrawChar(short x, short y, ushort col, CHARDEF* def);
-void PrintString(ushort x, ushort y, uchar col, const char* string, ushort flags);
+long GetStringLength(const char* string, long* top, long* bottom);
+void DrawChar(long x, long y, ushort col, CHARDEF* def);
+void PrintString(long x, long y, uchar col, const char* string, ushort flags);
 
 extern long stash_font_height;
 extern long smol_font_height;
